build: make the asciidoctor task repeatable safely
The asciidoctor sources are not all parts of the sourceDir of the task. So if we modify one of the .adoc files and rerun the task without cleaning first, the task is marked as up-to-date. This fixes the issue by adding all the .adoc files as inputs of the task.