Add contracts for testing WebServer module implementations
This helps to avoid copying and pasting tests for various web-server modules and to
introduce a way to enfore contract on all the implementations.

Any web-server module needs to implement WebServerModuleTest base class and
override single property server. The test suite is then located within
the base class and changing the base class will automatically run the
specified tests on all the implementations.