Updated download links on samples README page
The download links were broken on the samples page because the spring-fu-version attribute in the top level README file was not being read.  I am not sure if there is a way to achieve this on Github.  Although this solution is not ideal, I propose adding a separate attribute for the spring-fu-version on the samples page so that the download links are functional.