Type-safe customization of an existing application
The goal of this commit is to allow the configuration of an existing
application with type-safe configuration properties instead of using
error prone environment variables like now.

One of the design challenge is to be able to use configuration
properties in the DSL directly which is tricky because currently
they are beans and the context is not refreshed yet.

The proposed solution is to change
`inline fun <reified T : Any> configurationProperties(prefix: String = "")`
to
`inline fun <reified T : Any> configurationProperties(properties: T? = null, prefix: String = ""): T`
and to use `fun` instead of `val` for the application and the configuration.

Likely to be refined, but this is a major Kofu principle.

Closes gh-228